changes to take advantage of new(?) stuff
Don't think I really want/need that layer 3, first off. So, get rid of that.

LED 3 is now used to indicate that flashing is primed with the SYSFL combo.

The key that used to toggle L3 will now be used to do screen lock. On layer 0 it will send the Windows (and usual Ubuntu-ish) combo of GUI-L. On layer 1 it will send the macOS combo of CTRL-GUI-Q.

ESC is replaced by GESC ("grave escape"). Normally this does ESC; use GUI modifier to do the ` keycode or shift modifier to do ~ keycode.

The key that used to have grave/tilde will now toggle "caps word" behavior. (All caps until next space.)